pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
dependencies = [
"warp-lang",
"trimesh",
"pyglet",
"torch",
]
[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.tasks]
start = "python test.py"
